+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Commerce Changelog
 - Added the [`pdfPaperSize`](https://craftcommerce.com/docs/configuration#pdfpapersize) config setting.
 - Added the [`pdfPaperOrientation`](https://craftcommerce.com/docs/configuration#pdfpaperorientation) config setting.
 - Added a new setting to the Stripe Gateway to optionally include the [`receipt_email`](https://stripe.com/docs/api#create_charge-receipt_email) within request payment requests.
+- Added the [commerce_transactions.onCreateTransaction](https://craftcommerce.com/docs/events-reference#commerce_transactions.oncreatetransaction) event, which enables plugins to modify a newly created transaction model.
 
 ## 1.2.1342 - 2017-05-24